+++ changelog.md
@@ -1,6 +1,9 @@
 # changes
 
-## 5/15/15
+# 5/16/16
+ - Added an option for a human to play a computer (class `HumanAIGame`). Note that it's still a little buggy (for example it played 2 moves in 1 turn one time).
+
+## 5/15/16
 
  - Added a AI game - 2 computers that play eachother.
  - Refactored some methods. Changed some variables and methods to static (most notably `Cube.allowed`).